requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain_ollama
langchain-community>=0.3.0
streamlit>=1.40.0
python-dotenv>=1.0.0
duckduckgo-search>=6.0.0
langsmith>=0.1.0
plotly>=5.0.0
pandas>=2.0.0